+++ src/plugins/data_source_management/public/auth_registry/authentication_methods_registry.test.ts
@@ -3,22 +3,9 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import {
-  AuthenticationMethodRegistery,
-  AuthenticationMethod,
-} from './authentication_methods_registry';
+import { AuthenticationMethodRegistery } from './authentication_methods_registry';
 import React from 'react';
-
-export const createAuthenticationMethod = (
-  authMethod: Partial<AuthenticationMethod>
-): AuthenticationMethod => ({
-  name: 'unknown',
-  credentialForm: React.createElement('div', {}, 'Hello, world!'),
-  credentialSourceOption: {
-    value: 'unknown',
-  },
-  ...authMethod,
-});
+import { createAuthenticationMethod } from '../mocks';
 
 describe('AuthenticationMethodRegistery', () => {
   let registry: AuthenticationMethodRegistery;
