+++ src/plugins/data_source_management/public/auth_registry/index.ts
@@ -8,5 +8,3 @@ export {
   AuthenticationMethod,
   AuthenticationMethodRegistery,
 } from './authentication_methods_registry';
-
-export { createAuthenticationMethod } from './authentication_methods_registry.test';
